+++ .gitignore
@@ -1,0 +1,2 @@
+vendor/
+internal/config/config.json
