+++ .vscode/launch.json
@@ -1,0 +1,16 @@
+{
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Launch WhatsApp App",
+            "type": "go",
+            "request": "launch",
+            "mode": "auto",
+            "program": "${workspaceFolder}/src",
+            "env": {
+                "CGO_ENABLED": "1",
+                "GOARCH": "amd64"
+            }
+        }
+    ]
+}
