+++ go.mod
@@ -1,0 +1,26 @@
+module github.com/aaditya999/go-whatsapp-bot
+
+go 1.25.0
+
+require (
+	github.com/mattn/go-sqlite3 v1.14.32
+	go.mau.fi/whatsmeow v0.0.0-20250820160106-21f5124c7602
+)
+
+require (
+	filippo.io/edwards25519 v1.1.0 // indirect
+	github.com/google/uuid v1.6.0 // indirect
+	github.com/gorilla/websocket v1.5.0 // indirect
+	github.com/mattn/go-colorable v0.1.14 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/petermattis/goid v0.0.0-20250813065127-a731cc31b4fe // indirect
+	github.com/rs/zerolog v1.34.0 // indirect
+	go.mau.fi/libsignal v0.2.0 // indirect
+	go.mau.fi/util v0.9.0 // indirect
+	golang.org/x/crypto v0.41.0 // indirect
+	golang.org/x/exp v0.0.0-20250813145105-42675adae3e6 // indirect
+	golang.org/x/net v0.43.0 // indirect
+	golang.org/x/sys v0.35.0 // indirect
+	golang.org/x/text v0.28.0 // indirect
+	google.golang.org/protobuf v1.36.7 // indirect
+)
