--- src/.gitignore
@@ -1,2 +1,0 @@
-vendor/
-config.json
